src/backend/requirements.txt
transformers
fastapi
requests
pillow